rest/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
click==8.1.3
Flask==2.2.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
minio==7.1.12
nats-py==2.2.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
